backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "E-Commerce Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0"
]
